# app/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BarberBook API",
    "description": "API for barber appointment booking system",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/book": {
      "post": {
        "summary": "Create a new appointment",
        "description": "Books a new appointment with the provided details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/slots": {
      "get": {
        "summary": "Get available time slots",
        "description": "Retrieves available time slots for a specific date",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-06-15"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available slots retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/update": {
      "patch": {
        "summary": "Update an appointment",
        "description": "Updates, confirms, or cancels an existing appointment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an appointment",
        "description": "Permanently deletes an appointment",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing appointment ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointments": {
      "get": {
        "summary": "Get all appointments",
        "description": "Retrieves all appointments (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}": {
      "patch": {
        "summary": "Update appointment status",
        "description": "Updates the status of a specific appointment (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete appointment",
        "description": "Permanently deletes a specific appointment (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "BookingRequest": {
        "type": "object",
        "required": ["name", "phone", "service", "date", "time"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "pattern": "^[\\+]?[1-9][\\d]{0,15}$",
            "example": "+1234567890"
          },
          "service": {
            "type": "string",
            "enum": ["haircut", "beard", "both"],
            "example": "haircut"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-06-15"
          },
          "time": {
            "type": "string",
            "example": "10:00"
          }
        }
      },
      "UpdateRequest": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "example": "abc123"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled"],
            "example": "confirmed"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "pattern": "^[\\+]?[1-9][\\d]{0,15}$",
            "example": "+1234567890"
          },
          "service": {
            "type": "string",
            "enum": ["haircut", "beard", "both"],
            "example": "haircut"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-06-15"
          },
          "time": {
            "type": "string",
            "example": "10:00"
          }
        }
      },
      "BookingResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "appointmentId": {
            "type": "string",
            "example": "abc123"
          },
          "message": {
            "type": "string",
            "example": "Appointment booked successfully"
          }
        }
      },
      "SlotsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-06-15"
          },
          "availableSlots": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["09:00", "09:30", "10:00", "14:00"]
          }
        }
      },
      "UpdateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Appointment updated successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid request data"
          },
          "details": {
            "type": "string",
            "example": "Name must be at least 2 characters"
          }
        }
      },
      "AppointmentsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          }
        }
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "abc123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "service": {
            "type": "string",
            "enum": ["haircut", "beard", "both"],
            "example": "haircut"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-06-15"
          },
          "time": {
            "type": "string",
            "example": "10:00"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled"],
            "example": "pending"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-06-15T08:30:00Z"
          }
        }
      },
      "StatusUpdateRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled"],
            "example": "confirmed"
          }
        }
      }
    }
  }
}